tagifyEmpty: fix deletion of incorrect events
The Python implementation uses repo.iterevents(types=Commit), which
returns each commit along with its index in the repository's event
array. The Go version uses repo.commits(), which returns a slice of
events but does not report the indices of those events. Rather than
fetching the event index of each commit returned by repo.commits(),
tagifyEmpty() incorrectly treats the commit's position in the returned
slice as if it is the commit's index in []repo.events, which is not at
all correct. Fix this.