Reposurgeon.parseExpression(): stop eating standalone '?'
A standalone '?' token is not itself a valid selection expression; it
must follow some other expression upon which to operate.
SelectionParser itself is careful to avoid consuming tokens which are
not part of the selection expression syntax, leaving any such tokens
alone and returning those trailing items to the caller. It is incorrect,
therefore, for Reposurgeon's parseExpression() override to consume a '?'
which is not preceded by some other expression since such a case does
not constitute valid selection expression syntax. Fix it to instead
leave a standalone '?' untouched.

A consequence of this fix -- of no longer swallowing a standalone '?' --
is that it restores behavior in which '?' is passed along to the command
interpreter where it is treated as an alias of "help".