fix matching of standalone legacy-ID's and tags
reposurgeon 3.5 added the convenience of matching legacy-ID's and tags
without having to surround them with <...>, however, that support has
long (or always) been broken for top-level selection set evaluation
since it feeds the entire input line to Repository.named(), rather than
just the first token, and then wraps the entire line in <...> if named()
returns true. So, for instance, given the command "some-tag resolve", it
would ask named() if "some-tag resolve" existed, which would always
fail. Even if it did somehow match, it would then wrap the entire line
in <...>, resulting in "<some-tag resolve>", which is an invalid
command.

The only case in which it could work correctly was for the few commands
which do some secondary selection set evaluation as part of their own
argument processing (such as "remove", "tag", etc.), and even then only
if that was the last argument on the line.

Fix legacy-ID and tag convenience matching so that it works in all cases.