test/parser.tst: limit unnecessarily noisy "expected" output
The here-doc test dumps the entire content of 'simple.fi' after applying
an authormap. This is very noisy and makes it difficult to manually
inspect test output when attempting to diagnose a problem. Since this
test isn't specifically about testing authormap functionality, but is
instead about testing whether here-doc functionality works, we can
achieve the same result without all the noise by applying the authormap
to a single event and inspecting just that event.