test/attribution: enable "relaxed" mode to verify error cases
In the Python version of reposurgeon, errors thrown by commands invoked
during a "script" run do not abort the script, but in the Go version
they do. Aborting the script is problematic for unit tests run via
"script" which, in addition to checking normal behavior, also check
error cases. To work around this problem, the "relax" command was added
to prevent errors from aborting scripts.