parseVisibility: quote garbled portion of line in error message
Quote the portion of the input about which parseVisibility() is
complaining in order to make it easily distinguishable, rather than
letting it run together with the text of the error message itself.

A secondary goal of quoting is so that the error message produced by the
Go version of reposurgeon exactly matches the one produced by the Python
version. This is important for unit tests which check this error case
since the error message needs to match the test's "expected" text.