evalPathsetFull: retrofit to use PathMap
In the Python version of reposurgeon, eval_pathset_full() employs
PathMap, however, PathMap was not available in the Go version at the
time evalPathsetFull() was implemented in Go, so an alternate
brute-force but simple mechanism was employed instead.

Now that PathMap is available in the Go port, retrofit evalPathsetFull()
to take advantage of it.