Address GitLab issue #129: TestHasReferences fails
One of the regular expressions for Subversion revisions got too loose
and broke a unit test. Bug report said:

The case that fails is {"svn", false, " 3.14159 "}. The reason is that
3.14159 is matching the regular expression [0-9]+\w, which is what got
added in 2b533414.

See for example here. It's matching 1415 to the [0-9]+ part of the
regular expression and the 9 to the \w part. Since the regular
expression is not anchored, it's skipping over the "3." prefix.

I've reverted the change that brole this.  I've also added gotest as a
prerequisite target for goregress so we wont get mugged this way again.