Supply a missing reset in deleteall.fi.
Without this reset the behavior of the Python and Go versions
diverges. It looks like either the Python and Go logic for writing
incremental resets are buggy in different ways, but it's hard to know
as when fast-export generates resets is not well documented.
Similatly, the Python logic for generating branch resets in Subversion
translation might be wrong too.

For now, go for consistent Python & Go behavior at the cost of
leavung weird cases untested.