Separate interactivity from logging level.
Twirly batons are emitted on;y when the session is interactive,
which is initialized by checking is stdin is a tty.  Verbose
level 1 no longer has a meaning.

All tests pass at this revision, though many check files needed to
be trivially rebuild because of changes in response echoing,
and test files needed set/clear interactive commands added.