Only remember the first clique of each type
Instead of putting all the trivial cliques at the end, only take care of
the first nontrivial one (and the trivial one too for compatibility with
a non-documented case). Since all the other cliques will be sorted by
branch name later, that's enough.

The cliqueBranches is still modified to keep the changes localized but a
future commit will update this.