Reduce cliqueBranches role to sorting split commits by branchname
Since the code no longer cares about the order in which the branch
cliques were encountered, stop recording that order. This enables
populating cliqueBranches at the very last moment, just to sort it by
name. In particular, no care is needed anymore to avoid the clique
assigned to the base commit since it will not be in cliqueBranches.