Dispatch an awkward squaddie - branchreplace.svn
This required a major change to the way branch copies are handled.
Creation of fileops on a branch copy is no longer suppressed.  It
all comes out in the wash because the git importer generally
ignores them.

There is one exception: the emptyfrom.svn test oad. See the FIXME
comment.