Reallocate a string in DoChangelogs/parseAttribution
If we do not reallocate, the string will use the storage of an
underlying string that came from the whole Changelog file, and the
GC will thus decide to keep that around. That uses far too much memory
for a single attribution. This changes gives a 100x reduction of memory
not freed after DoChangelogs on a small exerpt of the GCC repository.

Fixes issue #154