Better handling of attributions in DoChangelogs
Stray "<" and ">" that are not part of the email address are not
permitted in fast-import streams and would make them invalid, but
bailing out is too extreme. Instead, we can remove those spurious
characters to make the attribution valid.

Additionally, some of those "<" and ">" might come from legit parenthesis in
the name, that a previous massaging converted to angle brackets to cater
for old-style attributions where round brackets delimited emails.
In front of the actual email block where angle brackets should remain,
replace the brackets by round ones.