when we open a file with O_CREATE, we also want O_TRUNC
O_CREATE will create the file if it doesn't exist, but if it already
exists then we need to specify O_TRUNC or it won't be
truncated. Writes to the file will start at the beginning, but if we
write less data than the file already contains then the end of the
file will remain unchanged. Havoc will ensure.