all commands now treat a nil selection as their preferred default
Some commands operate on the whole repository, others bail with an
error message, etc.

We avoid setting rs.selection to any particular value, preferring to
interpret the nil with some meaning, but leave the nil in place. The
primary benefit at the moment is that we never set it to repo.all()
and then turn around and compare it with the result of calling
repo.all() again. That comparison is particularly slow, and DoWrite in
particular was suffering as a result.