Append ".1" to the legacy cookie of first split commits
When commits were split, the first commit's legacyID was correctly
updated with a ".1" suffix, but its "SVN:" cookie in the legacy map
was not. Delete the old entry, and insert a new one with the correct
cookie for consistency with the other split commits.

Also update svnLinkFixups which now needs to look for split commit
cookies, and improve the comments therein.