Compare with --nobranch when on the "root" branch
Reposurgeon puts content into the "root" branch if it lies outside the
branchify namespaces. In that case, we should compare against the root
of the repository. This is a stop gap, but will still fail if there is
other content (in the standard namespaces) alongside the "root" changes.
Such repositories are broken anyway and the user can always not check
such revisions, that are most of the times not useful to keep.