Don't sort operations before commit splitting
This can create wrong history where a new branch is split off in the
middle of file modifications and in the git space it would seem that it
is split either before or after. It can also loose content since
sortOperations() discards everything before the last deleteall, without
knowing that our deleteall operations are special and have a path.

It is safe to sort the operations in the next phase.