Replace explicit --notrunk by the smarter --accept-empty
The latter will fail with a non-empty diff if trunk/master is missing
only in one of the repositories, which is more correct. If both sides
have no trunk, it will do the right thing and succeed.

When the repository has no structure at all, still compare with
--nobranch because reposurgeon maps the "root" branch back to "master"
which repotool would try to compare against an empty SVN trunk.