Search for link base backwards and not forward
When finding branch root parents, search backwards from the fromRev
revision to find the latest commit that touched the branch. If we search
forward we might link to a commit that changes the tree compared to what
it was at the fromPath:fromRev. It did not impact the contents due to
the protection added by the deleteall but meant the link would be
non-intuitive.

This was only a problem when the target commit didn't touch the correct
branch, which is rare.