Move the mergeinfo phase just after link fixups
The mergeinfo phase relies on the legacy map and on lastRelevantCommit()
both of with are not updated to reflect commit deletions when tagifying
or deleting junk. The consequence is that it could consider a lot of links
to disconnected commits that are not anymore into the repository, which
will fail the topology-based merge cleanup algorithm.

Also, the latter phases like dead branches pruning actually need the
topology coming from merges to operate correctly.