Fix splitting SVN commits that start with non-branch change (issue 183).
This fixes issue 183, where an SVN commit that mixes changes on a
declared branch and changes not on any declared branch fails to get
split if the non-branch changes come before any branch changes in the
SVN dump for the revision.  The problem is the
"commit.fileops[j].Source != oldbranch" not matching in the case where
the first fileop has "" (no branch) as Source, matching the initial ""
value of oldbranch.  (If the provided test SVN dump is rearranged so
the no-branch changes come *after* the branch changes, the commit is
already split correctly.)  The fix is simply to always count the first
fileop as starting a new clique, even when on branch "".