Restore logic setting sp.branches.
Commit 0e63e2ef65175a0c6b525b1bd84e10cf510edae7 ("Rework link fixups
to never mess with the contents") removed the code in the SVN dump
reader that set sp.branches to track branch roots.  sp.branches is
used in svnProcessTagEmpties to determine the name to give a tag for
an empty commit; failing to set anything in sp.branches meant that
such tags had emptycommit-* names instead of using the original name
of the tag in SVN.  This patch restores the removed logic, placing it
in what seems the most appropriate place at the end of
svnProcessBranches.  This should fix the remaining cases of issue 177
(or at least the case from the included trivial test).