Complete fix for file properties changed with directory copy (issue 171).
For the new logic in seekAncestor to work correctly, nodes being
copied into expandedNodes must also be renumbered to have a single
sequence of indices, so use appendExpanded also for copied nodes.

Also change the logic in seekAncestor to use 1-based indices
consistently, so not skipping checking the first node in the list,
although in the case of the present bug the first node should not be
relevant (generated file copies come after original directory copy).