Re-add gitignore files on branch copyfrom
They would miss since most of the times there wouldn't be any explicit
file operation, only those created from the expansion of the directory
copies. This also avoids cases where a branch was created and an svn
ignore added before the first file operation. When a latter file
operation would be encountered, we detected that the branch still had no
default gitignore and overwrote the custom one if the latter was at
toplevel.