Use a hash to speed up seekAncestor.
The code in seekAncestor that looks back through every previous node
for the current commit has overhead quadratic in the number of nodes
in a commit, making it very slow for commits with many fileops.
Instead use a hash mapping paths to the last node in the current
commit with that path.

This reduced time for phase 5 in one GCC conversion run from 8h17m13s
to 1h33m56s (runs compared not using exactly the same reposurgeon
version and also running into swap, so not a perfect comparison).