Add new option to control keeping refs/deleted commits.
The logic in svnProcessJunk, both before and after nerfing, uses
--nobranch as a condition for not deleting commits on refs/deleted
branches.  (Because of the nerfing, in the absence of that option it
actually only deletes the tip deleteall commit, leaving the rest
behind, but the apparent intent would be to delete the ref and all
commits not reachable from some non-deleted ref.)

This condition on --nobranch doesn't make much sense, since with
--nobranch there are no tags or branches, and so should be no deleted
refs.  Keeping such refs can sometimes be useful, for validation of
the conversion if nothing else (I want to keep them for test
conversions of GCC to provide additional points at which to validate
tree contents).  So this patch adds a --preserve option to disable
this commit deletion, as a replacement for --nobranch in that context.

Note that the tests reflect the results generated by current newsvn3
reposurgeon.  That is, delref.chk includes the refs/deleted commits
that are currently generated and would need updating for a safe
unnerfing of this code in svnProcessJunk.