Detect more merges: accept ranges missing revs before spin-off
SVN tends to not put in mergeinfo the revisions that predate the merge
base of the source and dest branch tips. We can accept such ranges
starting after the merge base instead of deciding they are cherry picks.

Computing a merge base would be costly, but we can make a poor man's one
by using the parent of the dest branch root if that parent is on the
source.