call baton.Sync in Reposurgeion.PostCmd, because the prompt is not printed via the baton
Sometimes the prompt appears to be intermingled with output from some
command, particularly if that output has many lines. This is because
Kommandant prints the prompt to stdout rather than to to the
baton. (Also, the baton would need more work in order to do the prompt
itself.) Calling the Sync method ensures that Kommandant waits for the
baton to catch up.