the Baton generates log messages of its own, and they should get sent to the logfile, if any
When the Baton is in use and endCounter/endProcess/endProgress is
called, the final state of the status line is rendered and printed as
a log message. This commit provides a way for the those messages to be
redirected to the log file, if there is one, without unduely coupling
the Baton with the rest of reposurgeon.

fixes #203