Unconditionally use the legacyID in a message
The message can only be emitted from SVN processing anyway so checking
if the VCS source is really SVN doesn't make sense. Using the commit
mark here is less than helpful since a later renumbering phase will
modify a lot of those (and especially this one since the message is
about a commit that is being deleted).