Rename the branch "root" to "master" if the latter isn't used
When the repository uses the canonical namespaces structure, commits to
the repository root can be considered as mistakes, and should be moved
to a special "root" branch. If instead the structure is not used and
everything is committed in the repository root, there is no need to have
a separate "root" branch. Map back the "root" branch to "master" in that
case.

This is a convenience for very simple repositories that should have been
read with the --nobranch option. The latter will do the right thing in
case of flat repositories even if they happen to have a trunk subfolder.