Copy the property map to avoid deleting all svn:ignore props
To avoid propagating svn:ignore properties when copying, the code
deleted the property from the stashed properties. But the stashed map
was a pointer to the original map in the node, which meant that every
svn:ignore property was lost. Use a copy instead, and don't overwrite
the node props member with that copy.