Properly copy mergeinfo ranges when there is a gap.
The existing logic failed to produce a complete union of mergeinfo
ranges in cases where there were both gaps and overlaps, resulting in
a suboptimal set of merge parents for GCC commit 152464.  The
svnmerge-integrated property in the included test illustrates the
problem, although it's only like to appear in practice when merge
information from svnmerge-integrated and svn:mergeinfo is combined.