Refactor simplify()
Use a logic similar to manifests, but internal to the commit. This
should be much faster than doing as many O(n^2) passes as needed for the
result to be invariant.

It actually fixed a bug where an M followed by a D were both deleted,
but the D must stay because maybe the M overwrote an already existing
file, which would leak through if the D was removed too. See the changes
in the restructure.chk test.