Compare master against the root if there is no structure
If the repository has no structure, then reposurgeon maps back the
special root branch to master since the latter does not exist. Handle
that gracefully in liftcheck.

This will still fail if the repository had some structure in previous
revisions bus has none in the last. In that case liftcheck will
mistakenly think that master corresponds to root when that is not the
case.