Change no-forward-merge to really test a bogus case
Now that empty revisions have commits until tagification, the mergeinfo
pass will correctly create a merge commit (with no changes). Use a
really bogus test.

The second test is now a duplicate, and thus tests another thing (that
dir-only revisions can be the target of a merge).

wrong-mergeinfo needs to be changed in the same vein. It still tests a
mergeinfo whose revision is above the highest revision in the repo, so
is not a complete duplicate.