Don't allow the committer's timezone to leak into author entries.
Re GitLan issue #216, jsm28 said:

Suppose that first an author map is read with the "authors" command,
giving a commit a timezone for both author and committer dates because
the author found through the map has a timezone there, then the author
of that commit is changed via the "changelogs" command to a different
attribution that does not have a timezone listed in the author map or
determined from their email address. I'd expect that the timezone for
the author date to default to UTC, but instead it defaults to the
committer's timezone.