Do not output "from" if possible for implicit parents
If omitting the "from" instruction still gives the correct parent, and
that parent was implicit when reading the input fast-import stream
(there was no "from"), then do not output the "from" line. This makes
such streams round-trip correctly.

Let the user disable that feature if she wants unconditionally explicit
parents. Also add a round-trip test with implicit parents, and a test
for disabling the implicit export. These tests are based on reference
so for now implicit-to-explicit.chk is identical to reference.chk.

Closes #223.