Avoid choosing a branch at random in deleteBranch
When choosing the branch to rewrite all commit.Branch members into, the
code used the first good candidate. But since the order of commit
children is not stable and depends on the order in which parent links
are created, this could induce spurious test failures.

Select the greatest good candidate in lexicographic order (so that the
empty value does not win).