Accept merges omitting the branch root revision
Before "svn merge" could detect itself which revisions were merged, it
was customary to specify -r <branch-root-revision>:<revision-to-merge>
when merging, where <branch-root-revision> is the revision creating
the source branch.

SVN, in its great wisdom, interpreted -r x:y as « branch is already
merged up to x, add a range from x+1 to y included », which skipped
the branch root revision.

Since this is a common case caused by a quirk of SVN, accept it too
instead of rejecting it as a cherry pick.

Closes #228