Remove stale entries in the legacy map after a squash
And do so in writeLegacyMap too, instead of just skipping them.
There can be quite a lot of them if the detetion set was big; this is
the case when pruning deleted SVN branches when --preserve is off. As
long as they are in the legacy map, the deleted commits will survive.