Improve walkManifests to drop manifests on childless commits
Before, only when processing an inheriting child could the manifests be
forgotten. Now drop the manifest as soon as the hook has returned, if we
know no other commit will need the manifest to inherit from (and stop the
recursion).

Also add a test, checking if out of the 279 commits of implicit.fi only
6 manifests at most are remembered simultaneously (they correspond to
interleaved commits or branches forking points, for which we have to
keep manifests a bit longer in memory), and that we correctly clean up
at the end of the method.