Attempt to back out the bug in Gitlab issue #226.
Edward Cree said: As far as I can tell, the three-directory monte in
rebuildRepo is broken and has been since "Correct file preservation -
.git config is precious!": the .git (or equivalent) that we just
populated in the staging directory (through git fast-import) is not
copied from staging to target, apparently to avoid clobbering
.git/config, hook scripts etc. But this means that we have either no
.git at all (in @shlomif's case, where target was an empty dir), or a
.git holding the old repository history, though combined with a
checkout of the new repository.

We move back to a simpler rename-based shuffle, counting on putting
config and hook files into the presrtve set to do the right thing.