Avoid O(n^2) costs when filling in the paths of a commit
As Joseph Myers discovered, we use an orderedStringSet(), which looks in
the existing entries to avoid duplicates. This can incur O(n^2) costs
w.r.t the number of touched paths in the commit, which can become very
big for branch copies on repositories with 100000 files.

Use a map to avoid duplicates, and convert to an orderedStringSet only
at the end.