Replace the markToEvent() cache by markToIndex()
In some cases, the markToEvent() cache was being stale, and code like

if x := repo.markToEvent(mark); x != nil {
    // do something with x
}

would fail, while

if idx := repo.markToIndex(mark); idx != -1 {
    x := repo.events[idx]
    // do something with x
}

would work correctly. Also, there are so many users of markToIndex()
that the probability of the cache not being rebuilt is very low. That
means the markToEvent() cache, even if fixed, might be an over
optimization.