Fill the markToIndex cache incrementally
Instead of waiting for a few queries before populating the cache
completely, always fill it up to the found event, and remember the len
of the prefix event slice whose cache is correct, to restart searching
and filling the cache from there.

A failed search will populate the complete cache.