Add a missing cache invalidation
Repository.tagifyEmpty() uses the mark to index cache internally, thus
needs it to be valid. RepositoryList.expunge() mutates the event list
without invalidating the cache, which is wrong. This currently works
because tagifyEmpty adds an event which clears the cache, before using
it, but we intend to clear the cache less and that would uncover the bug.