Use a big enough capacity for lastCommitOnBranchAt
lastCommitOnBranchAt[branch] is a list from revision numbers to commits.
When revisions are contiguous and start at 0, the maximum revision number is
len(sp.revisions)-1 and a list length of len(sp.revisions) is enough.

If either of these assumptions is wrong, then the maximum revision can
be higher and we need to allocate more. Use the correct value.