Remove unneeded sequence mutation declarations
Now that repo.addEvent() does not need to invalidate the markToIndex
cache, declaring a sequence mutation would trigger such an invalidation
and increase the computation cost for no reason.

The second thing declareSequenceMutation invalidates is the assignment
list, but at that point the user cannot have created assignments since
we are importing a new SVN repository.