Avoid invalidating needlessly when setting a mark
When calling setMark() on a (Blob or Commit) event whose mark was
previously empty, we could not know whether the event was in the
currently cached part of the event stream, but was skipped due to an
empty mark. This forced us to invalidate the cache just in case.

But that situation is very rare: most times, if not all the time, the
mark is empty because we are setting it just prior or just after adding
the event to the stream. In such a case invalidating the cache is going
to waste all previous caching for no reason.

To handle such cases without introducing false positives, remember if
while caching we skipped an event with an empty mark *and* a setMark
method. If so, we still need to invalidate the cache, because that could
very well be the one we are handling, but in all other cases, keep the
cache intact.