make the 'profile start' and 'profile save' commands symmetric
Some profile types start automatically and must be saved explicitly,
while others must be started explicitly but then are automatically
saved. The user shouldn't have to care about this, so we should hide
the difference.

To that end, 'profile start' stores the filename for later use when
the profile is started automatically. 'profile save' uses either the
filename specified on the command line or the one stored for later. At
program exit, we save any profile for which we have a stored filename.

fixes #261