Don't remove last character when undoing byte-stuffing
When byte stuffed '.a line\n' becomes '..a line\n'. When undoing this it
became '.a line' losing the '\n'. This could leave commit messages
without a trailing linefeed, something that reposurgeon objects to. It
could also corrupt a blob being edited if the file happens to contain
lines beginning with '.' as might happen in a .gitigmore.