Cache subtree gitHashes if possible
When a subtree pathmap is shared, it is immutable and we can avoid
computing its gitHash again and again. When it is not shared, we cannot
now for sure that the contents will not be changed, invalidating any
cached hash. The likelyhood of needing the hash of an unshared pathmap
multiple times is low anyway, so we need not bother creating a clever
invalidation scheme.