Remember destination paths of copies, not source
When saving the copied subtree of a copy node for later expansion,
reposurgeon used to save it at its source path, later rewriting the
paths when expanding nodes. Directly remember the correct path instead.

This is simpler, avoids several string allocations, and sidesteps a
crash that could occur when the source path is not long enough, in case
a file exists with the same name as the source directory (this can only
happen in a corrupt SVN dump, but letting reposurgeon crash is not
very polite).