Reduce the weight of blobs, assuming that $-cookies are rare.
Each blob used to have an inline Cookie structure to hold the parsed
value of an $-cookie found in it.  Now it contains a pointer to a
Cookie object on the heap if it found and paarsed a $-cookie, nil
otherwise.

A heap pointer is 8 bytes, a Cookie 32 (string storage used is the
same for both).  The new representation pays 8 more bytes of overhead
per actual cookie.  Thus the new reprentation will consume more total
storage if and only if more than 75% of blobs have $-cookies.

Use of $-cookies has fallen over time so we should expect the new
represntation to give better results at ime passes.