Rip out retention of expunged content in a synthetic repository.
This festurectomy is a response to a long string of bugs and
troublesome edge cases culminating in GitLab issue #274: Reposurgeon
crashes when expunging from a freshly read git repo. I now think this
was just too difficult to get right.

The new ~ operator of expunge should fill in most of the missing
functionality.  If it doesn't, I'll add a repository clone operation.

As a bonus, this change reduces the size of Blob and Commit objects.