Use Golang's library temporary-directory primitive.
Demonstrating once again that the purpose of shipping major
stable releases is so the Dead God Finagle can tap you on the
shoulder and say "Gotcha!".

This fixes a theoretical security bug.  Very theoretical; the attacker
would already have to have your credentials or root on your machine to
exploit it, would have to no of a compromising failure in a repository
translation, and would have to intervene while repotool is running a
comparison in order to mask the failure by substituting the
source-repository version of some file(s).

Anyone who could do this would probably find it easier to subvert
your copy of repotool to turn off mismatch detection entirely.