Change the workflow in the standard conversion makefile to eliminate a pass.
We now go direct from stream dump of the local mirror (which is often a
Subversion dump) to live repository.  This meand there will be only
one reposurgeon read of a strem equivalent to the entire history, not
two.  Which is good, because for larhe repositories those reads
totally dominate conversion time.

The old workflow made more sense when the Subversion dump reader was
still much slower than it is now.  Under it, surgical moves could be
tried on the second-stage stream dump without having to reinterpret
the first one.  But it meant that you would always pay the time
cost for two full stream reads, which is bad on very large repositories.

Now, you only pay that cost once.